+++ inc/admin/options.php
@@ -1385,7 +1385,77 @@ function dli_register_main_options_metabox() {
 	);
 
 	/**
-	* 11 - Registers options page "Progetti".
+	* 11 - Registers options page "Spin-off".
+	*/
+	// Intestazione della sezione.
+	$args = array(
+		'id'           => 'dli_options_spinoff',
+		'title'        => esc_html__( 'Le Spin-off', 'design_laboratori_italia' ),
+		'object_types' => array( 'options-page' ),
+		'option_key'   => 'spinoff',
+		'parent_slug'  => 'dli_options',
+		'capability'   => DLI_EDIT_CONFIG_PERMISSION,
+		'tab_group'    => 'dli_options',
+		'tab_title'    => __( 'Spin-off', 'design_laboratori_italia' ),
+	);
+	// 'tab_group' property is supported in > 2.4.0.
+	if ( version_compare( CMB2_VERSION, '2.4.0' ) ) {
+		$args['display_cb'] = 'dli_options_display_with_tabs';
+	}
+	$spinoff_options = new_cmb2_box( $args );
+	$spinoff_landing_url = dli_get_template_page_url( 'page-templates/spinoff.php' );
+	$descr = sprintf( __( 'Inserisci qui le informazioni utili a popolare <a href="%s"> la pagina di panoramica delle Spin-off</a>', 'design_laboratori_italia' ), $spinoff_landing_url	);
+	$spinoff_options->add_field(
+		array(
+		'id'   => $prefix . 'spinoff_istruzioni',
+		'name' => __( 'Sezione Le Spin-off', 'design_laboratori_italia' ),
+		'desc' => $descr,
+		'type' => 'title',
+		)
+	);
+	
+	// Campi descrizione della sezione.
+	$spinoff_options->add_field(
+		array(
+			'id'         => $prefix . 'testo_spinoff',
+			'name'       => __( 'Descrizione Sezione', 'design_laboratori_italia' ),
+			'desc'       => __( 'es: "Spin-off del Laboratorio."' , 'design_laboratori_italia' ),
+			'type' => 'wysiwyg',
+			'options' => array(
+				'textarea_rows' => 1,
+				'media_buttons' => false,
+				'teeny'         => true,
+				'quicktags'     => false,
+				'tinymce'       => array(
+					'toolbar1'       => 'bold,italic,link,unlink,undo,redo',
+					'valid_elements' => 'a[href],strong,em,p,br', 
+				),
+			),
+		)
+	);
+
+	$spinoff_options->add_field(
+		array(
+			'id'         => $prefix . 'testo_spinoff'. DLI_ENG_SUFFIX_LANGUAGE,
+			'name'       => __( 'Descrizione Sezione ENG', 'design_laboratori_italia' ),
+			'desc'       => __( 'es: "Spin-off of the Lab"' , 'design_laboratori_italia' ),
+			'type' => 'wysiwyg',
+			'options' => array(
+				'textarea_rows' => 1,
+				'media_buttons' => false,
+				'teeny'         => true,
+				'quicktags'     => false,
+				'tinymce'       => array(
+					'toolbar1'       => 'bold,italic,link,unlink,undo,redo',
+					'valid_elements' => 'a[href],strong,em,p,br', 
+				),
+			),
+		)
+	);
+
+
+	/**
+	* 12 - Registers options page "Progetti".
 	*/
 	// Intestazione della sezione.
 	$args = array(
@@ -1526,7 +1596,7 @@ function dli_register_main_options_metabox() {
 	);
 
 	/**
-	* 12 - Registers options page "Attività di ricerca".
+	* 13 - Registers options page "Attività di ricerca".
 	*/
 	$args = array(
 		'id'           => 'dli_options_ricerca',
@@ -1598,7 +1668,7 @@ function dli_register_main_options_metabox() {
 
 
 	/**
-	* 13 - Registers options page "Attività di ricerca".
+	* 14 - Registers options page "Attività di ricerca".
 	*/
 		$args = array(
 				'id'           => 'dli_options_luoghi',
@@ -1672,7 +1742,7 @@ function dli_register_main_options_metabox() {
 		));
 
 	/**
-	* 14 - Registers options page "Social media".
+	* 15 - Registers options page "Social media".
 	*/
 		$args = array(
 				'id'           => 'dli_options_socials',
@@ -1770,7 +1840,7 @@ function dli_register_main_options_metabox() {
 	// BEGIN SECTION FOR ADMINISTRATORS
 	if ( current_user_can( DLI_ADMIN_EDIT_CONFIG_PERMISSION ) ) {
 		/**
-		* 15 - Registers options page "Integrazione con Indico".
+		* 16 - Registers options page "Integrazione con Indico".
 		*/
 		$args = array(
 			'id'           => 'dli_options_indico',
@@ -1976,7 +2046,7 @@ function dli_register_main_options_metabox() {
 		);
 
 		/**
-		* 16 - Registers options page "Integrazione con IRIS".
+		* 17 - Registers options page "Integrazione con IRIS".
 		*/
 		$args = array(
 			'id'           => 'dli_options_iris',
@@ -2128,7 +2198,7 @@ function dli_register_main_options_metabox() {
 
 
 		/**
-		* 17 - Registers options page "Altro".
+		* 18 - Registers options page "Altro".
 		*/
 
 		$args = array(